wip(dui): experimental new stream flow
