+++ DesktopUI2/DesktopUI2/Controls/SendButton.xaml
@@ -1,0 +1,58 @@
+<UserControl
+  x:Class="DesktopUI2.Controls.SendButton"
+  xmlns="https://github.com/avaloniaui"
+  xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+  xmlns:assists="clr-namespace:Material.Styles.Assists;assembly=Material.Styles"
+  xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+  xmlns:icons="clr-namespace:Material.Icons.Avalonia;assembly=Material.Icons.Avalonia"
+  xmlns:m="clr-namespace:Material.Styles;assembly=Material.Styles"
+  xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+  mc:Ignorable="d">
+  <Grid Margin="4">
+    <m:FloatingButton
+      VerticalAlignment="Center"
+      Command="{Binding SendCommand}"
+      IsDefault="True"
+      IsEnabled="{Binding Progress.IsProgressing, Converter={x:Static BoolConverters.Not}}"
+      ToolTip.Tip="Sends to this stream">
+      <Button.Content>
+        <icons:MaterialIcon
+          Width="20"
+          Height="20"
+          Foreground="White"
+          Kind="CubeSend">
+          <icons:MaterialIcon.RenderTransform>
+            <RotateTransform Angle="-30" />
+          </icons:MaterialIcon.RenderTransform>
+        </icons:MaterialIcon>
+      </Button.Content>
+    </m:FloatingButton>
+
+    <Grid>
+      <ProgressBar
+        Width="40"
+        Height="40"
+        Classes="Circle"
+        IsIndeterminate="{Binding Progress.IsIndeterminate, FallbackValue=True}"
+        IsVisible="{Binding Progress.IsProgressing, FallbackValue=False}"
+        Maximum="{Binding Progress.Max}"
+        Value="{Binding Progress.Value}" />
+      <m:FloatingButton
+        assists:ButtonAssist.HoverColor="Transparent"
+        assists:ShadowAssist.ShadowDepth="0"
+        Background="Transparent"
+        Classes="Mini"
+        Command="{Binding CancelSendOrReceiveCommand}"
+        IsVisible="{Binding Progress.IsProgressing, FallbackValue=False}"
+        ToolTip.Tip="Cancel">
+        <Button.Content>
+          <icons:MaterialIcon
+            Width="15"
+            Height="15"
+            Foreground="{DynamicResource PrimaryHueDarkBrush}"
+            Kind="Close" />
+        </Button.Content>
+      </m:FloatingButton>
+    </Grid>
+  </Grid>
+</UserControl>
