+++ DesktopUI2/DesktopUI2/Controls/SendButton.xaml.cs
@@ -1,0 +1,19 @@
+using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace DesktopUI2.Controls
+{
+  public partial class SendButton : UserControl
+  {
+    public SendButton()
+    {
+      InitializeComponent();
+    }
+
+    private void InitializeComponent()
+    {
+      AvaloniaXamlLoader.Load(this);
+    }
+  }
+}
