+++ DesktopUI2/DesktopUI2/Models/Filters/AllSelectionFilter.cs
@@ -19,7 +19,7 @@ namespace DesktopUI2.Models.Filters
     {
       get
       {
-        return "Everything";
+        return "everything";
       }
     }
   }
