+++ DesktopUI2/DesktopUI2/Styles/Styles.xaml
@@ -8,4 +8,5 @@
   <StyleInclude Source="/Styles/Dialog.xaml" />
   <StyleInclude Source="/Styles/Text.xaml" />
   <StyleInclude Source="/Styles/TextBox.xaml" />
+  <StyleInclude Source="/Styles/Expander.xaml" />
 </Styles>
